Reorganise dask predict into own module
